requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==3.2.13
django-appconf==1.0.5
django-compressor==3.1
django-heroku==0.3.1
django-libsass==0.9
django-widget-tweaks==1.4.11
djangorestframework==3.13.1
gunicorn==20.1.0
importlib-resources==5.4.0
libsass==0.21.0
# pkg-resources==0.0.0
psycopg==3.0.11
psycopg2==2.9.3
pytz==2022.1
rcssmin==1.1.0
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.2
typing-extensions==4.1.1
whitenoise==5.3.0
zipp==3.6.0
